Added testcases to reveal bug
Discovered a bug where html files with names ending in index.html are having index.html stripped from the url incorrectly.  For example, aindex.html becomes just "a".  Bug is due to code that excludes index.html from url if the filename is index.html.

This commit includes new test cases to reveal this bug.  The build should fail when this commite is pushed (to confirm test cases before pushing the fix).